sapphire: Import Parts from fuxi
fuxi: Enable touchscreen feature based on settings

Single tap gesture and always on udfps rely on mode 16 to be enabled, so
setup a listener for those settings and set the mode enabled if one of
both is on

fuxi: Start parts services as early as possible on bootup

BOOT_COMPLETED intent is broadcast only after the device is unlocked and
user data (encrypted storage) is available. Use LOCKED_BOOT_COMPLETED
intent for services that don't require accessing data.